solved 56.merge-intervals.

+++ problems/56.merge-intervals.hpp
@@ -6,14 +6,33 @@
 #include <array>
 #include <set>
 
+/**
+ * Given an array of intervals where intervals[i] = [starti, endi], merge all overlapping intervals, and return an array of the non-overlapping intervals that cover all the intervals in the input.
+ *
+ * Example 1:
+ * Input: intervals = [[1,3],[2,6],[8,10],[15,18]]
+ * Output: [[1,6],[8,10],[15,18]]
+ * Explanation: Since intervals [1,3] and [2,6] overlaps, merge them into [1,6].
+ *
+ * Example 2:
+ * Input: intervals = [[1,4],[4,5]]
+ * Output: [[1,5]]
+ * Explanation: Intervals [1,4] and [4,5] are considered overlapping.
+ *
+ * Constraints:
+ * 1 <= intervals.length <= 104
+ * intervals[i].length == 2
+ * 0 <= starti <= endi <= 104
+*/
+
 class Solution {
 public:
   std::vector<std::vector<int>> merge(const std::vector<std::vector<int>>& intervals) {
     std::vector<std::vector<int>> sln;
 
     const int MAX_END_VALUE = 10000;
-    std::array<int, MAX_END_VALUE + 1> xAxis{};
-    std::set<int> oneBoundInterval;
+    std::array<int16_t, MAX_END_VALUE + 1> xAxis{};
+    std::set<int16_t> oneBoundInterval;
 
     int begin = intervals[0][0];
     int end = intervals[0][0];
